Batch process FRET data

This function applies average_technical_replicates and correct_fret_signal to all raw data files present in a given directory, and either writes output files in another directory or returns a list containing corrected datasets for programmatic manipulation.

batch_process(input = NULL, titrations = NULL, blanks = NULL,
  skip_lines = 4, output = NULL)

Arguments

input
A directory containing raw data files, or a named list of already loaded datasets (as dataframes).

titrations
A character vector containing words that identify the titration series (e.g. titration). These words must match the ones in raw_data$Content (case sensitive).

blanks
A character vector containing words that identify the blank experiment series (e.g. blank). These words must match the ones in raw_data$Content (case sensitive).

skip_lines
Number of lines to skip at the beginning of CSV files (see read.csv). Defaults to 4.

output
An optional directory where to write output files.

Value

Write to directory_out data files containing the corrected FRET signal from each input raw data file. Each output file contains two columns named concentration and fret_corrected, and has the same file name as the original raw data file appended with _corrected. If no output directory is specified, the function instead returns a list containing the corrected datasets, which can then be accessed programmatically.

See also

average_technical_replicates and correct_fret_signal, which can be used manually on an individual dataset.