Detect which binding model equation was used to fit a dataset

This internal function takes a model object from fit_binding_model or nls and returns the name of the binding model used to fit the dataset.

detect_binding_model(fit)

Arguments

fit
An nls object. The output of fit_binding_model can be used directly as input here.

Value

A character vector corresponding to the name of the binding model equation used to fit the dataset corresponding to the model object.