Fit a binding model equation to the experimental FRET data

This function uses initial guesses of kd, fret_min and fret_max to fit a binding model equation to the corrected experimental FRET data. This function is a wrapper around nls and provides reasonnable default settings (i.e. refine only kd, fret_min and fret_max, and possibly the Hill coefficient n, but keep probe concentration fixed).

fit_binding_model(corrected_data, binding_model = NULL, parameters,
  donor_concentration = 10)

Arguments

corrected_data
A dataframe containing the corrected FRET signal. It must contain at least two columns named fret_corrected and concentration. The output of the correct_fret_signal function can be used directly as input here.

binding_model
A binding model equation to fit to the experimental data. Possible values are hyperbola, quadratic and quadratic_homodimer.

parameters
A named list containing initial guesses for kd, fret_min and fret_max. The output of guess_parameters can be used directly as input here.

donor_concentration
The concentration of donor-labeled molecule. Defaults to 10.

Value

An named list containing: fit, an nls object containing the results of the fit; binding_model, the equation used in the fitting procedure; and donor_concentration, the donor concentration used in the fitting procedure. This list can be directly used as input to make_figure.