Calculate fluorescence polarization and anisotropy from raw data

This function calculates fluorescence polarization and anisotropy values from the raw data of parallel and perpendicularly polarized channels.

fp_calculate_pola_aniso_int(raw_data)

Arguments

raw_data
A dataframe containing the raw data. It must contain at least 7 columns named:

Experiment

A unique name identifying each experiment.

Type

For example, "titration".

Replicate

A number identifying the technical replicate (1, 2, etc.).

Observation

A number identifying each observation in a titration series (corresponds to the plate column numbers, if experiments are set up as rows in a 384-well plate). The number of observations for an experiment and its blanks must match, and a given observation number must associate data points at the same concentration in the titration series.

concentration

The ligand concentration in the titration series.

parallel

The name of the column containing fluorescence intensities from the parallel channel.

perpendicular

The name of the column containing fluorescence intensities from the perpendicular channel.

The output of fp_format_data can be used directly as input for this function.

Value

A dataframe containing the same columns as the input dataframe, plus three columns called polarization, anisotropy and intensity.

See also

fp_format_data to prepare datasets for use with fp_calculate_pola_aniso_int.