Correct FRET signal

This function corrects the raw FRET signal by applying corrections for donor bleedthrough and acceptor direct excitation. See Hieb AR et al (2012) and Winkler DD et al (2012) for details.

fret_correct_signal(data, output_directory = NULL)

Arguments

data
A dataframe containing the FRET data after avergaing over technical replicates. This dataframe must contain the following columns:

Experiment

A unique name identifying each experiment.

Type

Either "blank" (no donor) or "titration" (donor present).

Observation

A number identifying each observation in a titration series (corresponds to the plate column numbers, if experiments are set up as rows in a 384-well plate). The number of observations for an experiment and its blanks must match, and a given observation number must associate data points at the same concentration in the titration series.

concentration

The ligand concentration in the titration series.

fret_channel

Fluorescence intensity in the FRET channel.

acceptor_channel

Fluorescence intensity in the acceptor channel.

donor_channel

Fluorescence intensity in the donor channel.

The output of fret_average_replicates can be used directly as input for this function.

output_directory
An optional output directory name where to write corrected data in CSV files. This directory will be created if it does not already exist.

Value

A dataframe containing the corrected FRET signal. It contains three columns:

Experiment

A unique name identifying each experiment (same as in the input data).

concentration

The ligand concentration in the titration series (same as in the input data).

signal

The corrected FRET signal.

See also

fret_average_replicates to prepare a dataset for use with fret_correct_signal.

For details on the signal correction applied by fret_correct_signal, see:

Hieb AR et al (2012) Fluorescence Strategies for High-Throughput Quantification of Protein Interactions. Nucleic Acids Research 40 (5): e33 (doi:10.1093/nar/gkr1045) and

Winkler DD et al (2012) Quantifying Chromatin-Associated Interactions: The HI-FI System. In Methods in Enzymology pp 243–274. Elsevier (doi:10.1016/B978-0-12-391940-3.00011-1).