Format raw experimental FRET binding data for subsequent processing

This function formats data for subsequent processing by fret_average_replicates. It will accept all data files in a given directory, or any number of data files specified by their names, or any number of dataframes supplied as a named list, or even a single dataframe supplied as is.

fret_format_data(input = NULL, skip_lines = 0, metadata_json = NULL)

Arguments

input
A directory name, a vector of file names, a single dataset already loaded in memory as a dataframe, or a list of datasets already loaded in memory as dataframes.

skip_lines
The number of lines to skip at the beginning of CSV files (usually containing header information before the actual data starts).

metadata_json
A JSON file describing mappings between internal names and metadata information in the actual data files. It can contain the following entries:

content

The name of the column describing sample content. Default value: "Content".

concentration

The name of the column containing the concentration series. Default value: "concentration".

fret_channel

The name of the column containing fluorescence intensities from the FRET channel. Default value: "fret_channel".

acceptor_channel

The name of the column containing fluorescence intensities from the acceptor channel. Default value: "acceptor_channel".

donor_channel

The name of the column containing fluorescence intensities from the donor channel. Default value: "donor_channel".

titration

The name describing points of the titration series. Default value: "titration".

acceptor_only

The name describing points of the control series with only acceptor. Default value: "acceptor_only".

donor_only

The name describing points of the control series with only donor. Default value: "donor_only".

buffer_only

The name describing points of the control series with only buffer. Default value: "buffer_only".

Value

A single dataframe with the combined input data, containing 8 columns: Experiment, Type, Replicate, Observation, fret_channel, acceptor_channel, donor_channel and concentration. Rows with missing values will be dropped.