Inspect raw fluorescence data

This function plots the raw fluorescence data from the donor, acceptor and FRET channels, as a function of the acceptor concentration (titration series).

inspect_raw_data(raw_data, titrations = NULL, highest_signal = NULL)

Arguments

raw_data
A dataframe containing the raw fluorescence data. It must contain at least five columns named Content, donor_channel, acceptor_channel, fret_channel and concentration. The Content column is used to plot distinct data series with different symbols; it must contain, for each row, a word describing which data series or replicate this row belongs to (like "blank_1", "titration_1", etc.).

titrations
An optional character vector containing the names associated to the titration series. If not specified, the donor channel plot will not display the mean and +/- 10 % shaded area.

highest_signal
An optional number corresponding to the maximal signal measurable by the plate reader instrument used. Defaults to NULL, which won't check for the presence of saturated reads. The input number is not checked in any way: make sure it really corresponds to a saturated read with your instrument.

Value

A list containing a logical value `saturated_reads` indicating the presence of saturated reads in the dataset (equals to `TRUE` if saturated reads are present, `FALSE` if there is no saturated read, or `NULL` if this was not tested), and three ggplot2 graph objects (named donor, acceptor and fret). Warning messages appear when missing values are encountered, and can be safely ignored.

Examples

not_run({
  # Send all plots to the output device.
  inspect_raw_data(fret_good)
  
  # Donor plot will display pipetting precision.
  inspect_raw_data(fret_good, c("titration_1", "titration_2")
  
  # Store plots in a variable, print all plots, print only one of them.
  my_plots <- inspect_raw_data(fret_good)
  my_plots
  my_plots$acceptor
})

#> Error: <text>:9:3: unexpected symbol
#> 8:   # Store plots in a variable, print all plots, print only one of them.
#> 9:   my_plots
#>      ^